[backend] allow comment update only for owner of the comment

+++ backend/src/main/java/dev/lschen/cookit/comment/CommentService.java
@@ -2,11 +2,14 @@ package dev.lschen.cookit.comment;
 
 import dev.lschen.cookit.recipe.Recipe;
 import dev.lschen.cookit.recipe.RecipeService;
+import dev.lschen.cookit.user.User;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -30,14 +33,22 @@ public class CommentService {
                 .orElseThrow(() -> new EntityNotFoundException("Comment not found"));
     }
 
-    public Comment patchById(CommentRequest request, Long commentId) {
+    public Comment patchById(CommentRequest request, Long commentId, Authentication authentication) {
         Comment comment = findById(commentId);
+        User user = (User) authentication.getPrincipal();
+        if (!Objects.equals(user.getUsername(), comment.getCommentedBy().getUsername())) {
+            throw new RuntimeException("Cannot update other users comment");
+        }
         comment.setContent(request.content());
         return commentRepository.save(comment);
     }
 
-    public void deleteById(Long commentId) {
-        findById(commentId);
+    public void deleteById(Long commentId, Authentication authentication) {
+        Comment comment = findById(commentId);
+        User user = (User) authentication.getPrincipal();
+        if (!Objects.equals(user.getUsername(), comment.getCommentedBy().getUsername())) {
+            throw new RuntimeException("Cannot delete other users comment");
+        }
         commentRepository.deleteById(commentId);
     }
